requirements.txt
# Core Streamlit and Frontend
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
pillow>=10.0.0
numpy>=1.24.0

# Backend and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0

# Computer Vision (simplified)
opencv-python>=4.8.0

# Audio Processing (simplified)
SpeechRecognition>=3.10.0

# Additional Utilities
python-dateutil>=2.8.2
pydantic>=2.5.0 